Add fixtures and tests for projections and strategies
- Introduced new fixtures for `Projection` objects in `conftest.py` for testing projections.
- Updated tests for projections to utilize the new fixtures and simplified the test structure.
- Added new fixtures for `Strategy` objects in `conftest.py` for testing strategies.
- Refactored strategy tests to use the new fixtures and improved test coverage for impact and hazard measures.
- Cleaned up unnecessary code and comments in the strategy tests.
- Updated imports in the scenario runner tests for consistency.
