+++ tests/test_adapter/test_sfincs_adapter.py
@@ -2,7 +2,6 @@ import platform
 import tempfile
 from copy import copy
 from datetime import datetime, timedelta
-from functools import partial
 from pathlib import Path
 from typing import Tuple
 from unittest import mock
@@ -12,7 +11,6 @@ import numpy as np
 import pandas as pd
 import pytest
 import xarray as xr
-from cht_cyclones.tropical_cyclone import TropicalCyclone
 
 from flood_adapt.adapter.sfincs_adapter import SfincsAdapter
 from flood_adapt.config.hazard import (
@@ -22,51 +20,40 @@ from flood_adapt.config.hazard import (
     RiverModel,
     WaterlevelReferenceModel,
 )
-from flood_adapt.dbs_classes.database import Database
+from flood_adapt.dbs_classes import Database
 from flood_adapt.dbs_classes.interface.database import IDatabase
-from flood_adapt.objects.events.hurricane import TranslationModel
-from flood_adapt.objects.events.synthetic import (
+from flood_adapt.objects import (
+    FloodWall,
+    GreenInfrastructure,
+    Measure,
+    Projection,
+    Pump,
+    Scenario,
+    SelectionType,
     SyntheticEvent,
 )
-from flood_adapt.objects.forcing import unit_system as us
-from flood_adapt.objects.forcing.discharge import (
+from flood_adapt.objects.events.hurricane import TranslationModel
+from flood_adapt.objects.forcing import (
     DischargeConstant,
     DischargeCSV,
     DischargeSynthetic,
-)
-from flood_adapt.objects.forcing.forcing import (
-    ForcingSource,
     ForcingType,
-    IDischarge,
     IForcing,
-    IRainfall,
-    IWaterlevel,
-    IWind,
-)
-from flood_adapt.objects.forcing.rainfall import (
     RainfallConstant,
     RainfallCSV,
     RainfallMeteo,
     RainfallNetCDF,
     RainfallSynthetic,
     RainfallTrack,
-)
-from flood_adapt.objects.forcing.time_frame import (
-    TimeFrame,
-)
-from flood_adapt.objects.forcing.timeseries import (
     ShapeType,
-    TimeseriesFactory,
-)
-from flood_adapt.objects.forcing.waterlevels import (
     SurgeModel,
     TideModel,
+    TimeFrame,
+    TimeseriesFactory,
     WaterlevelCSV,
     WaterlevelGauged,
     WaterlevelModel,
     WaterlevelSynthetic,
-)
-from flood_adapt.objects.forcing.wind import (
     WindConstant,
     WindCSV,
     WindMeteo,
@@ -74,281 +61,12 @@ from flood_adapt.objects.forcing.wind import (
     WindSynthetic,
     WindTrack,
 )
-from flood_adapt.objects.measures.measures import (
-    FloodWall,
-    GreenInfrastructure,
-    Measure,
-    Pump,
-    SelectionType,
-)
-from flood_adapt.objects.projections.projections import Projection
-from flood_adapt.objects.scenarios.scenarios import Scenario
-from tests.fixtures import TEST_DATA_DIR
+from flood_adapt.objects.forcing import unit_system as us
+from tests.test_adapter.conftest import _unsupported_forcing_source
 from tests.test_objects.test_forcing.test_netcdf import (
     get_test_dataset,
     time_model_2_hr_timestep,
 )
-
-
-@pytest.fixture()
-def default_sfincs_adapter(test_db) -> SfincsAdapter:
-    overland_path = test_db.static_path / "templates" / "overland"
-    with SfincsAdapter(model_root=overland_path) as adapter:
-        start_time = datetime(2023, 1, 1, 0, 0, 0)
-        duration = timedelta(hours=3)
-
-        adapter.set_timing(
-            TimeFrame(
-                start_time=start_time,
-                end_time=start_time + duration,
-                time_step=timedelta(hours=1),
-            )
-        )
-        adapter._ensure_no_existing_forcings()
-
-        return adapter
-
-
-@pytest.fixture()
-def sfincs_adapter_with_dummy_scn(default_sfincs_adapter):
-    # Mock scenario to get a rainfall multiplier
-    dummy_scn = mock.Mock()
-    dummy_event = mock.Mock()
-    dummy_event.rainfall_multiplier = 2
-    dummy_event.time = TimeFrame()
-    dummy_scn.event = dummy_event
-    default_sfincs_adapter._scenario = dummy_scn
-
-    yield default_sfincs_adapter
-
-
-@pytest.fixture()
-def sfincs_adapter_2_rivers(test_db: IDatabase) -> tuple[IDatabase, SfincsAdapter]:
-    overland_2_rivers = test_db.static_path / "templates" / "overland_2_rivers"
-
-    rivers = []
-    with open(overland_2_rivers / "sfincs.src", "r") as f:
-        lines = f.readlines()
-        for i, line in enumerate(lines):
-            x, y = line.split()
-            x = float(x)
-            y = float(y)
-            rivers.append(
-                RiverModel(
-                    name=f"river_{x}_{y}",
-                    mean_discharge=us.UnitfulDischarge(
-                        value=5000 + i * 100, units=us.UnitTypesDischarge.cms
-                    ),
-                    x_coordinate=x,
-                    y_coordinate=y,
-                )
-            )
-    test_db.site.sfincs.river = rivers
-
-    with SfincsAdapter(model_root=(overland_2_rivers)) as adapter:
-        adapter.set_timing(TimeFrame())
-        adapter._ensure_no_existing_forcings()
-
-        return adapter, test_db
-
-
-@pytest.fixture()
-def synthetic_discharge():
-    if river := Database().site.sfincs.river:
-        return DischargeSynthetic(
-            river=river[0],
-            timeseries=TimeseriesFactory.from_args(
-                shape_type=ShapeType.triangle,
-                duration=us.UnitfulTime(value=3, units=us.UnitTypesTime.hours),
-                peak_time=us.UnitfulTime(value=1, units=us.UnitTypesTime.hours),
-                peak_value=us.UnitfulDischarge(
-                    value=10, units=us.UnitTypesDischarge.cms
-                ),
-            ),
-        )
-
-
-@pytest.fixture()
-def test_river() -> RiverModel:
-    return RiverModel(
-        name="test_river",
-        mean_discharge=us.UnitfulDischarge(value=0, units=us.UnitTypesDischarge.cms),
-        x_coordinate=0,
-        y_coordinate=0,
-    )
-
-
-@pytest.fixture()
-def river_in_db() -> RiverModel:
-    return Database().site.sfincs.river[0]
-
-
-@pytest.fixture()
-def synthetic_rainfall():
-    return RainfallSynthetic(
-        timeseries=TimeseriesFactory.from_args(
-            shape_type=ShapeType.triangle,
-            duration=us.UnitfulTime(value=3, units=us.UnitTypesTime.hours),
-            peak_time=us.UnitfulTime(value=1, units=us.UnitTypesTime.hours),
-            peak_value=us.UnitfulIntensity(value=10, units=us.UnitTypesIntensity.mm_hr),
-        )
-    )
-
-
-@pytest.fixture()
-def synthetic_wind():
-    return WindSynthetic(
-        magnitude=TimeseriesFactory.from_args(
-            shape_type=ShapeType.triangle,
-            duration=us.UnitfulTime(value=1, units=us.UnitTypesTime.days),
-            peak_time=us.UnitfulTime(value=2, units=us.UnitTypesTime.hours),
-            peak_value=us.UnitfulVelocity(value=1, units=us.UnitTypesVelocity.mps),
-        ),
-        direction=TimeseriesFactory.from_args(
-            shape_type=ShapeType.triangle,
-            duration=us.UnitfulTime(value=1, units=us.UnitTypesTime.days),
-            peak_time=us.UnitfulTime(value=2, units=us.UnitTypesTime.hours),
-            peak_value=us.UnitfulDirection(
-                value=1, units=us.UnitTypesDirection.degrees
-            ),
-        ),
-    )
-
-
-@pytest.fixture()
-def synthetic_waterlevels():
-    return WaterlevelSynthetic(
-        surge=SurgeModel(
-            timeseries=TimeseriesFactory.from_args(
-                shape_type=ShapeType.triangle,
-                duration=us.UnitfulTime(value=1, units=us.UnitTypesTime.days),
-                peak_time=us.UnitfulTime(value=8, units=us.UnitTypesTime.hours),
-                peak_value=us.UnitfulLength(value=1, units=us.UnitTypesLength.meters),
-            )
-        ),
-        tide=TideModel(
-            harmonic_amplitude=us.UnitfulLength(
-                value=1, units=us.UnitTypesLength.meters
-            ),
-            harmonic_period=us.UnitfulTime(value=12.42, units=us.UnitTypesTime.hours),
-            harmonic_phase=us.UnitfulTime(value=0, units=us.UnitTypesTime.hours),
-        ),
-    )
-
-
-def _mock_meteohandler_read(
-    time: TimeFrame,
-    test_db: IDatabase,
-    *args,
-    **kwargs,
-) -> xr.Dataset | xr.DataArray:
-    gen = np.random.default_rng(42)
-    lat = [test_db.site.lat - 10, test_db.site.lat + 10]
-    lon = [test_db.site.lon - 10, test_db.site.lon + 10]
-    _time = pd.date_range(
-        start=time.start_time,
-        end=time.end_time,
-        freq=timedelta(hours=1),
-        name="time",
-    )
-
-    ds = xr.Dataset(
-        data_vars={
-            "wind10_u": (("time", "lat", "lon"), gen.random((len(_time), 2, 2))),
-            "wind10_v": (("time", "lat", "lon"), gen.random((len(_time), 2, 2))),
-            "press_msl": (("time", "lat", "lon"), gen.random((len(_time), 2, 2))),
-            "precip": (("time", "lat", "lon"), gen.random((len(_time), 2, 2))),
-        },
-        coords={
-            "lat": lat,
-            "lon": lon,
-            "time": _time,
-        },
-        attrs={
-            "crs": 4326,
-        },
-    )
-    ds.raster.set_crs(4326)
-
-    # Convert the longitude to -180 to 180 to match hydromt-sfincs
-    if ds["lon"].min() > 180:
-        ds["lon"] = ds["lon"] - 360
-
-    return ds
-
-
-@pytest.fixture()
-def mock_meteohandler_read(test_db):
-    with mock.patch(
-        "flood_adapt.adapter.sfincs_adapter.MeteoHandler.read",
-        side_effect=partial(_mock_meteohandler_read, test_db=test_db),
-    ):
-        yield
-
-
-@pytest.fixture()
-def mock_offshorehandler_get_resulting_waterlevels():
-    with mock.patch(
-        "flood_adapt.adapter.sfincs_offshore.OffshoreSfincsHandler.get_resulting_waterlevels"
-    ) as mock_get_data_wl_from_model:
-        df = pd.DataFrame(
-            data={"waterlevel": [1, 2, 3]},
-            index=pd.date_range("2023-01-01", periods=3, freq="H"),
-        )
-        mock_get_data_wl_from_model.return_value = df
-        yield df
-
-
-def _unsupported_forcing_source(type: ForcingType):
-    mock_source: ForcingSource = mock.Mock(
-        spec=(ForcingSource, str), return_value="unsupported_source"
-    )
-    mock_source.lower = mock.Mock(return_value="unsupported_source")
-    mock_source.capitalize = mock.Mock(return_value="Unsupported_source")
-
-    match type:
-        case ForcingType.DISCHARGE:
-
-            class UnsupportedDischarge(IDischarge):
-                source: ForcingSource = mock_source
-                river: RiverModel = mock.Mock(spec=RiverModel)
-
-            unsupported = UnsupportedDischarge()
-        case ForcingType.RAINFALL:
-
-            class UnsupportedRainfall(IRainfall):
-                source: ForcingSource = mock_source
-
-            unsupported = UnsupportedRainfall()
-
-        case ForcingType.WATERLEVEL:
-
-            class UnsupportedWaterlevel(IWaterlevel):
-                source: ForcingSource = mock_source
-
-            unsupported = UnsupportedWaterlevel()
-        case ForcingType.WIND:
-
-            class UnsupportedWind(IWind):
-                source: ForcingSource = mock_source
-
-            unsupported = UnsupportedWind()
-        case _:
-            raise ValueError(f"Unsupported forcing type: {type}")
-    return unsupported
-
-
-@pytest.fixture()
-def spw_file() -> Path:
-    cyc_file = TEST_DATA_DIR / "IAN.cyc"
-    spw_file = TEST_DATA_DIR / "IAN.spw"
-    if spw_file.exists():
-        return spw_file
-    tc = TropicalCyclone()
-    tc.include_rainfall = True
-    tc.read_track(cyc_file, fmt="ddb_cyc")
-    tc.to_spiderweb(spw_file)
-    return spw_file
 
 
 class TestAddForcing:
@@ -426,10 +144,14 @@ class TestAddForcing:
             assert default_sfincs_adapter.wind is not None
 
         def test_add_forcing_wind_from_track_cyc(
-            self, test_db, tmp_path, default_sfincs_adapter: SfincsAdapter
+            self,
+            test_data_dir,
+            test_db,
+            tmp_path,
+            default_sfincs_adapter: SfincsAdapter,
         ):
             # Arrange
-            track_file = TEST_DATA_DIR / "IAN.cyc"
+            track_file = test_data_dir / "IAN.cyc"
             default_sfincs_adapter._event = mock.Mock()
             default_sfincs_adapter._event.hurricane_translation = TranslationModel(
                 eastwest_translation=us.UnitfulLength(
@@ -473,8 +195,7 @@ class TestAddForcing:
         ):
             # Arrange
             tmp_path = Path(tempfile.gettempdir()) / "wind.csv"
-            t0, t1 = default_sfincs_adapter._model.get_model_time()
-            time_frame = TimeFrame(start_time=t0, end_time=t1)
+            time_frame = default_sfincs_adapter.get_model_time()
             synthetic_wind.to_dataframe(time_frame).to_csv(tmp_path)
 
             forcing = WindCSV(path=tmp_path)
@@ -589,10 +310,10 @@ class TestAddForcing:
             assert adapter.rainfall is not None
 
         def test_add_forcing_rainfall_track_cyc(
-            self, test_db, default_sfincs_adapter: SfincsAdapter
+            self, test_data_dir, test_db, default_sfincs_adapter: SfincsAdapter
         ):
             # Arrange
-            track_file = TEST_DATA_DIR / "IAN.cyc"
+            track_file = test_data_dir / "IAN.cyc"
             default_sfincs_adapter._event = mock.Mock()
             default_sfincs_adapter._event.hurricane_translation = TranslationModel(
                 eastwest_translation=us.UnitfulLength(
@@ -984,20 +705,19 @@ class TestAddMeasure:
     @staticmethod
     def get_measure_gdf(adapter: SfincsAdapter, measure: Measure) -> gpd.GeoDataFrame:
         measure = adapter.database.measures.get(measure.name)
-        measure_gdf = gpd.read_file(
-            adapter.database.measures.input_path / measure.name / measure.polygon_file
-        )
-        return measure_gdf.to_crs(adapter._model.crs.to_epsg())
+        if measure.gdf is None or not isinstance(measure.gdf, gpd.GeoDataFrame):
+            raise ValueError("Measure does not have a valid gdf attribute.")
+        return measure.gdf.to_crs(adapter._model.crs.to_epsg())
 
     class TestFloodwall:
         @pytest.fixture()
-        def floodwall(self, test_db) -> FloodWall:
+        def floodwall(self, test_data_dir, test_db) -> FloodWall:
             floodwall = FloodWall(
                 name="test_seawall",
                 description="seawall",
                 selection_type=SelectionType.polyline,
                 elevation=us.UnitfulLength(value=12, units=us.UnitTypesLength.feet),
-                polygon_file=str(TEST_DATA_DIR / "seawall.geojson"),
+                gdf=gpd.read_file(test_data_dir / "seawall.geojson").to_crs(epsg=4326),
             )
 
             test_db.measures.save(floodwall)
@@ -1083,7 +803,7 @@ class TestAddMeasure:
 
     class TestPump:
         @pytest.fixture()
-        def pump(self, test_db) -> Pump:
+        def pump(self, test_db, test_data_dir) -> Pump:
             pump = Pump(
                 name="test_pump",
                 description="pump",
@@ -1091,7 +811,7 @@ class TestAddMeasure:
                     value=100, units=us.UnitTypesDischarge.cfs
                 ),
                 selection_type=SelectionType.polyline,
-                polygon_file=str(TEST_DATA_DIR / "pump.geojson"),
+                gdf=gpd.read_file(test_data_dir / "pump.geojson").to_crs(epsg=4326),
             )
             test_db.measures.save(pump)
             return pump
@@ -1192,12 +912,14 @@ class TestAddMeasure:
 
     class TestGreenInfrastructure:
         @pytest.fixture()
-        def water_square(self, test_db) -> GreenInfrastructure:
+        def water_square(self, test_db, test_data_dir) -> GreenInfrastructure:
             green_infra = GreenInfrastructure(
                 name="test_greeninfra",
                 description="greeninfra",
                 selection_type=SelectionType.polygon,
-                polygon_file=str(TEST_DATA_DIR / "green_infra.geojson"),
+                gdf=gpd.read_file(test_data_dir / "green_infra.geojson").to_crs(
+                    epsg=4326
+                ),
                 volume=us.UnitfulVolume(value=1, units=us.UnitTypesVolume.m3),
                 height=us.UnitfulHeight(value=2, units=us.UnitTypesLength.meters),
                 percent_area=0.5,
